requirements.txt
Django==2.0.3,
django-shopping-cart
